requirements.txt
APScheduler==3.6.3
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
bitarray==1.2.2
certifi==2020.12.5
chardet==4.0.0
cytoolz==0.11.0
decorator==4.4.2
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
etherscan-py==0.2.3
hexbytes==0.2.1
idna==2.10
ipfshttpclient==0.7.0a1
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
jsonschema==3.2.0
lru-dict==1.1.7
multiaddr==0.0.9
mypy==0.812
mypy-extensions==0.4.3
netaddr==0.8.0
parsimonious==0.8.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.16
protobuf==3.15.3
ptyprocess==0.7.0
pycryptodome==3.10.1
Pygments==2.8.0
pyrsistent==0.17.3
python-dotenv==0.15.0
python-telegram-bot==13.3
pytz==2021.1
requests==2.25.1
rlp==2.0.1
schedule==1.0.0
six==1.15.0
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.3
varint==1.0.2
wcwidth==0.2.5
web3==5.17.0
websockets==8.1
